GamePlayFix

+++ 2024_winapigamep_framework_22/TimeManager.h
@@ -25,7 +25,7 @@ private:
     LARGE_INTEGER m_llPrevCnt = {};
     LARGE_INTEGER m_llCurCnt = {};
     LARGE_INTEGER m_llFrequency = {};
-    float         m_dT = 0.f;
+    double         m_dT = 0.f;
 
     // FPS(Frame Per Second)
     UINT          m_fps = 0;
